+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [9.9.7] - (20.03.2025)
+
+## Changes
+
+- added possibility to set custom encharge name displayed in the home app, closes [#188](https://github.com/grzegorz914/homebridge-enphase-envoy/issues/188)
+- added possibility to disable/display encharge light bulb accessory in the home app
+- bump dependencies
+- config schema updated
+- redme updated
+- cleanup
+
 ## [9.9.4] - (19.03.2025)
 
 ## Changes